backend/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3
sqlparse==0.3.1
